pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kavak-ai-agent"
version = "0.1.0"
description = "Kavak AI Sales Agent for WhatsApp"
authors = [
    {name = "Ferdinand Bracho"},
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.23.2",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "chromadb>=0.4.15",
    "redis>=5.0.1",
    "twilio>=8.10.0",
    "pydantic>=2.0.0",
    "pandas>=2.0.0",
    "sentence-transformers>=2.2.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
    "pytest>=7.4.0",
    "pydantic-settings>=2.9.1",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pytest-cov>=4.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
]
